#518 - more migration to multi tenant schema (memberType, jobType, event job, some of job)

+++ src/database/eventJob.ts
@@ -10,20 +10,20 @@ import { getEventList } from './event';
 import { getJobType } from './jobType';
 import { calculateStartDate } from '../util/dateHelper';
 
-export const GET_EVENT_JOB_SQL = 'SELECT * FROM v_event_job WHERE event_job_id = ?';
-export const INSERT_EVENT_JOB_SQL = 'INSERT INTO event_job (event_type_id, job_type_id, count) VALUES (?, ?, ?)';
-export const PATCH_EVENT_JOB_SQL = 'CALL sp_patch_event_job(?, ?, ?, ?)';
-export const DELETE_EVENT_JOB_SQL = 'DELETE FROM event_job WHERE event_job_id = ?';
+export const GET_EVENT_JOB_SQL = 'SELECT * FROM v_event_job WHERE event_job_id = ? AND tenant_id = ?';
 
 export async function insertEventJob(req: PostNewEventJobRequest, tenantId: string): Promise<number> {
     if (_.isEmpty(req)) {
         throw new Error('user input error');
     }
-    const values = [req.eventTypeId, req.jobTypeId, req.count];
+    const values = [req.eventTypeId, tenantId, req.jobTypeId, req.count];
 
     let result;
     try {
-        [result] = await getPool().query<OkPacket>(INSERT_EVENT_JOB_SQL, values);
+        [result] = await getPool().query<OkPacket>(
+            'INSERT INTO event_job (event_type_id, tenant_id, job_type_id, count) VALUES (?, ?, ?, ?)',
+            values,
+        );
     } catch (e: any) {
         if ('errno' in e) {
             switch (e.errno) {
@@ -45,7 +45,7 @@ export async function insertEventJob(req: PostNewEventJobRequest, tenantId: stri
     upcomingEvents.forEach(async (event) => {
         // now for each event, insert the job for that event
         // "2022-06-21 09:30:00"
-        const jobType = await getJobType(req.jobTypeId);
+        const jobType = await getJobType(req.jobTypeId, tenantId);
         const jobDate = calculateStartDate(event.start.toString(), jobType.jobDayNumber);
         const newJob: PostNewJobRequest = {
             eventId: event.eventId,
@@ -65,15 +65,15 @@ export async function insertEventJob(req: PostNewEventJobRequest, tenantId: stri
         // YAY for learning new things!
         const insertPromises = [];
         for (let index = 0; index < req.count; index++) {
-            insertPromises.push(insertJob(newJob));
+            insertPromises.push(insertJob(tenantId, newJob));
         }
         await Promise.all(insertPromises);
     });
     return result.insertId;
 }
 
-export async function getEventJob(id: number): Promise<EventJob> {
-    const values = [id];
+export async function getEventJob(id: number, tenantId: string): Promise<EventJob> {
+    const values = [id, tenantId];
 
     let results;
     try {
@@ -97,12 +97,14 @@ export async function getEventJob(id: number): Promise<EventJob> {
     };
 }
 
-export async function addJobTypeEvents(jobCountDiff: any, row: RowDataPacket, jobTypeId: number, modifiedBy: number) {
+// eslint-disable-next-line max-len
+export async function addJobTypeEvents(jobCountDiff: any, row: RowDataPacket, jobTypeId: number, modifiedBy: number, tenantId: string) {
     logger.info('Adding jobs for an event type as the count was changed.');
     for (let index = 0; index < jobCountDiff; index++) {
         const newJob: PostNewJobRequest = {
             eventId: row.event_id,
             jobTypeId,
+            tenantId,
             jobStartDate: calculateStartDate(row.start, row.job_day_number),
             jobEndDate: calculateStartDate(row.start, row.job_day_number),
             pointsAwarded: row.points_awarded,
@@ -114,7 +116,7 @@ export async function addJobTypeEvents(jobCountDiff: any, row: RowDataPacket, jo
         };
         try {
             // eslint-disable-next-line no-await-in-loop
-            const insertedJob = await insertJob(newJob);
+            const insertedJob = await insertJob(tenantId, newJob);
             logger.info(`inserted job with ID ${insertedJob} for event with id ${row.event_id}`);
         } catch (e: any) {
             logger.error(e);
@@ -123,15 +125,16 @@ export async function addJobTypeEvents(jobCountDiff: any, row: RowDataPacket, jo
     }
 }
 
-export async function removeJobTypeEvents(row: RowDataPacket, jobTypeId: number) {
+export async function removeJobTypeEvents(row: RowDataPacket, jobTypeId: number, tenantId: string) {
     logger.info('Removing jobs for an event type as the count was changed.');
     // bring some positivity in the world, no reason to be so negative!
     const removalCount = row.difference * -1;
     for (let index = 0; index < removalCount; index++) {
-        const deleteSql = 'delete from job where event_id = ? and job_type_id = ? and member_id is null limit 1';
+        const deleteSql =
+            'delete from job where event_id = ? and job_type_id = ? and member_id is null and tenant_id = ? limit 1';
         try {
             // eslint-disable-next-line no-await-in-loop
-            const [removeResults] = await getPool().query<OkPacket>(deleteSql, [row.event_id, jobTypeId]);
+            const [removeResults] = await getPool().query<OkPacket>(deleteSql, [row.event_id, jobTypeId, tenantId]);
             logger.info(`removed one job with id ${jobTypeId} for event id ${row.event_id}`);
             logger.info(removeResults.affectedRows);
         } catch (error: any) {
@@ -141,23 +144,25 @@ export async function removeJobTypeEvents(row: RowDataPacket, jobTypeId: number)
     }
 }
 
-export async function updateJobsOnEventJob(id: number, req: PatchEventJobRequest): Promise<number> {
+export async function updateJobsOnEventJob(id: number, tenantId: string, req: PatchEventJobRequest): Promise<number> {
     let results;
     let changedCount = 0;
     try {
         const countSql = `select distinct(event), event_id, job_type_id, start, end, points_awarded, cash_payout,
             meal_ticket, count(*) job_count, (?)-count(*) difference
-            from v_job where event_type_id = ? and job_type_id = ? and start > date_sub(now(), interval 7 day)
+            from v_job where tenant_id = ? and event_type_id = ? and job_type_id = ?
+            and start > date_sub(now(), interval 7 day)
             group by event`;
-        [results] = await getPool().query<RowDataPacket[]>(countSql, [req.count, req.eventTypeId, req.jobTypeId]);
+        const params = [req.count, tenantId, req.eventTypeId, req.jobTypeId];
+        [results] = await getPool().query<RowDataPacket[]>(countSql, params);
         results.forEach(async (row) => {
             const jobCountDiff = row.difference;
             if (jobCountDiff > 0) {
-                await addJobTypeEvents(jobCountDiff, row, req.jobTypeId as number, req.modifiedBy as number);
+                await addJobTypeEvents(jobCountDiff, row, req.jobTypeId as number, req.modifiedBy as number, tenantId);
             } else if (jobCountDiff < 0) {
-                await removeJobTypeEvents(row, req.jobTypeId as number);
+                await removeJobTypeEvents(row, req.jobTypeId as number, tenantId);
             } else {
-                logger.info('Called jobs upate but there was no count change so count update was not done.');
+                logger.info('Called jobs update but there was no count change so count update was not done.');
             }
             changedCount++;
         });
@@ -167,15 +172,17 @@ export async function updateJobsOnEventJob(id: number, req: PatchEventJobRequest
     return changedCount;
 }
 
-export async function patchEventJob(id: number, req: PatchEventJobRequest): Promise<void> {
+export async function patchEventJob(id: number, tenantId: string, req: PatchEventJobRequest): Promise<void> {
     if (_.isEmpty(req)) {
         throw new Error('user input error');
     }
-    const values = [id, req.eventTypeId, req.jobTypeId, req.count];
+    const values = [req.eventTypeId, req.jobTypeId, req.count, id, tenantId];
 
     let result;
     try {
-        [result] = await getPool().query<OkPacket>(PATCH_EVENT_JOB_SQL, values);
+        // eslint-disable-next-line max-len
+        const updateSql = 'update event_job set event_type_id = ?, job_type_id = ?, count = ? where event_job_id = ? and tenant_id = ?';
+        [result] = await getPool().query<OkPacket>(updateSql, values);
     } catch (e: any) {
         if ('errno' in e) {
             switch (e.errno) {
@@ -196,21 +203,25 @@ export async function patchEventJob(id: number, req: PatchEventJobRequest): Prom
     if (result.affectedRows < 1) {
         throw new Error('not found');
     }
-    const changedJobs = await updateJobsOnEventJob(id, req);
+    const changedJobs = await updateJobsOnEventJob(id, tenantId, req);
     logger.info(`Added or removed ${changedJobs} jobs while doing an update to eventJob`);
 }
 
-export async function deleteEventJob(id: number): Promise<void> {
-    const values = [id];
+export async function deleteEventJob(id: number, tenantId: string): Promise<void> {
+    const values = [id, tenantId];
 
     let result;
     try {
         // get the job right before deleting it so we know what it is. it's a
         // ghooooooooooost job.
-        const eventJob = await getEventJob(id);
-        [result] = await getPool().query<OkPacket>(DELETE_EVENT_JOB_SQL, values);
+        const eventJob = await getEventJob(id, tenantId);
+        [result] = await getPool().query<OkPacket>(
+            'delete from event_job where event_job_id = ? and tenant_id = ?',
+            values,
+        );
         const [deleteJobsResult] = await getPool()
-            .query<OkPacket>('delete from job where job_type_id = ? and job_start_date > now()', [eventJob.jobTypeId]);
+            // eslint-disable-next-line max-len
+            .query<OkPacket>('delete from job where tenant_id = ? and job_type_id = ? and job_start_date > now()', [tenantId, eventJob.jobTypeId]);
         logger.info(`removed job id ${id} from jobs table`);
         logger.info(`${deleteJobsResult.affectedRows} rows in jobs affected`);
     } catch (e) {
